Change the handling of not found wrapped typedefs
If a typedef target could not be resolved at module load time, now a
Python warning is generated instead of printing the error to stderr. The
content of the message was improved. It now contains information
about the missing C++ type.

The advantage of warnings is that they can be turned off or raise 
errors. This allows for tests on missing typedef attributes.

For Python warnings see:
https://docs.python.org/2/library/warnings.html

Change-Id: I610daaeea7c36395a15430047084d76bbd7015e4